+++ include/JSystem/JGadget/linklist.h
@@ -128,8 +128,8 @@ struct TLinkList : TNodeLinkList {
 
         iterator& operator=(const iterator& rhs) {
             //TODO: Probably fakematch? Not sure what's going on here exactly
-            (TIterator<std::bidirectional_iterator_tag, T, long, T*, T&>&)*this =
-                (const TIterator<std::bidirectional_iterator_tag, T, long, T*, T&>&)rhs;
+            (TIterator<std::bidirectional_iterator_tag, T, s32, T*, T&>&)*this =
+                (const TIterator<std::bidirectional_iterator_tag, T, s32, T*, T&>&)rhs;
             this->node = rhs.node;
             return *this;
         }
